Add Business to db
Fixes: #10

I got the Business table up and running, and hooked up the controller, routes, views, etc to make new ones and view persisted ones.

My current plan is to have the Business model function as the top of the hierarchy for the main models in this app. For example, in a cascading fashion, the layout would look like either: Business -> has Sites -> each Site has Members...... OR    Business -> has Site and Members. I think I will probably go with the latter, assuming membership would transfer between sites if somebody had a membership for a business.