Add some CSS
Fixes: #21

I'm beginning to add some CSS to the website to make it pretty, but I really suck at it, so it is slow going. It's tempting to use chat gpt to cook up the visuals for me, but then I lose understanding of what each piece is doing which defeats the purpose of this project. I am currently trying to figure out how to lay things out with flexbox, and trying to determine a good color scheme.