Add logout button
Fixes: #28

I added a logout button to kill a session and redirect to the login page. For some reason some JS packages weren't loading properly, and I couldn't get them to work, so the method from the blog wasn't working to sign me out. It relied on 'UJS' and UJS wasn't working.

This is the real way that isn't working:

link_to "Sign out", logout_path, method:"delete"

This is the hacky fix:

<%= form_with url: logout_path, method: :delete do %>
  <%= submit_tag "Sign out" %>
<% end %>

The form submission works because it is bypassing the UJS entirely, as it is submitting a POST request with a 'delete' method attached, and has a respective route in routes.rb to map it to the sessions#destroy action.
I still need to fix the UJS issue though.