Add Member table
Fixes: #9

Adds a member table and spec file for it. Members will be the real life customers that are subscribed to a business and visit a business' sites. For example, I am a member at Crunch fitness, and I visit their site on 23rd street. I need to fix this spec file though.