Add edit page to business
Fixes: #47

Adding an edit page for businesses so that they can customize their own settings. The first available customizations will be name and if they want to display a schedule or not.

For including a schedule, I've added the simple_calendar gem to the gemfile.

There is still more work to be done. The form submission for update doesn't actually update the model yet, and the calendar needs work, but it is at least present when a business has display_schedule set to true. I was considering adding a whole new model called business_config.rb to house configurations such as display_schedule, but I figured at this point the repo is so small that adding that complexity would be more of a hindrance than it would be helpful.