Small edits to business #create and #index
Add the necessary fields to the Business #create form in order to actually save a new valid business. Also rearranged a block of code to make the entire <div> disappear where the schedule would appear on Business #show if :display_schedule is turned off. Previously an empty white box would still be there, but now the entire thing is hidden.