Reorder some of the links
I rearranged some of the event related redirects to be more consistent. For example, any 'back to business' link should be the furthest left. Any 'new/edit event' links should be furthest right. Just a subtle thing to make UX less clunky, though the UI is so bad that it probably doesn't matter.