Add event specs
I've added both the model and request specs for Event. The model specs can be filled out further, and the request specs aren't working. Something is going on where the server is blocking the requestand returning a 403. I need to do a whole PR of adding request specs and furthering the model specs on everything. Especially businesses, now that they can have a schedule with events.