Add interests
Now events and members can be associated through interests. A member can designate which events interest them and with a status such as the default: 'Interested'. I'm still working on adding methods for setting these easily, and the ability to do this through the Gymothy website is nonexistent. I only have it working in the console. I'll probably have to make a whole separate SSA admin portal (me), admin portal (business owner), and member portal (members of businesses). Within the member portal there can be a button to display interest.