Integrate Stripe payment processing
Fixes: #56

I've integrated Stripe with Gymothy using Stripe's Free Test API keys. I was shocked at how simple it was to get this running. I had to add a .env and .env-example file in order to include the API keys (.env is in .gitignore). I've also included the 'dotenv' and 'stripe' gem. I will need to add specs to test this behavior better, and research further into the API itself to see its full use cases, but I'm happy that it seems to be working. Well, working from the CLI. The next step will be to implement the UI.