Add Rails CSS template
Fixes: #57

I've downloaded a CSS template from [bootstrap.com](https://startbootstrap.com/themes) and attempted to use this [Rails/Bootstrap log post](https://matthieubertrand5.medium.com/how-to-use-a-bootstrap-theme-in-a-ruby-on-rails-project-2a24f01c6f44) as a guide for integrating it into Gymothy. I've ran into a few issues and decided to come back to it later. I may instead attempt to do my own custom CSS upon completion of the Codecademy CSS course, so I'm putting this on hold.