Edit _nav_dropdown partial to allow form_links, make memberships unique to member/business pairs, build forms for site#new and membership#new
I created a form for a business to Add Members. It takes you to a form that is secretly building a Membership with a list of existing Members as options. I will probably have to add the ability to create a new Member if necessary.

I created a similar form for a business to Add Site.

I made Memberships uniquely indexed on :business_id and :member_id so there isn't multiple instances of the same pairing.

I edited the _nav_dropdown partial to include fancy form_submission links, so now Sites, Members, and Employees can all use the partial in Business#show.

Made a bunch of new views for Membership, and a new controller for it.

Updated Site views.

Next I want to use JS to listen 'onclick' for the actual nav bar main buttons to take you to the index pages for those items.