Fix member_spec.rb visit spec, add links to visit#show
I fixed the pending spec in member_spec.rb about tracking the count of visits a member makes to a site. The visit just needed to be reloaded before doing the final expect. Also, I added links to thesite and member show pages from an associated Visit#show. Previously, only the business was linked, and the site and member name were displayed as plain text.