Fix :count increment on Visit#create
When creating a Visit with the UI, you use a _form. The visit method for a member instance finds or creates a new visit when called, and then increments the :count by 1. So when the DB creates a new visit instance with a default of 0, it immediately gets set to 1, IF you're using the .visit method. When creating a visit instance with a form, the visit method doesn't get called, so the :count is displayed as 0 which doesn't make any sense IRL. I fixed this by adding a .visit call to the member instance that's associated with the site after the visit is successfully saved on Visit#create.