Begin adding management page for a member to manage its memberships
Fixes: #62

Before adding the Stripe UI, I need a page for a member to view and manage all of their memberships. I started making a table style to build a good looking table, and am working on a _table partial that I can simply pass my content, title, etc to and have a table render on any page. The partial needs more work to be generic, but is currently working for our use case. I need to add a membership#show and #edit view as well, to view and edit specific memberships.