Hide businesses in URL (renamed to b) and fix some css
I don't like how long these URLs are becoming, so I'm going to periodically start changing the routes' paths to read to something simpler, as I did for businesses->b. It makes the URL cleaner. I also want to hide the :id numbers for objects, so I'll probably slugify things like member name and stuff like that.

I also fixed some CSS that was bothering me. The <content> tag for the _table partial would spill the title out of its box if the window was too small. Now it doesn't.