Generalize _table partial
I'm hype about this one. Now I can create tables with basically whatever content I want using the same partial. Here's an example call:

render partial: 'partials/table', locals: {
                                  title: 'Memberships',
                                  headers: ['Business', 'Status', 'Subscription'],
                                  content: @memberships,
                                  attributes: ['business.name', 'status', 'subscription'],
                                  btn_name: 'Manage',
                                  btn_link: -> (m) { business_membership_path(m.business, m) },
                                }

Basically it builds the table using the title and headers and each row will correspond to a membership from a loop in the view iterating on locals[:content]. The attributes are what we want to access within that content, so in this case we want to do something like membership.business.name in the first column. That's where the tricky/cool part comes in. Since the attrs are passed as a list of strings, we split them on '.' and then use inject in combination with public_send to build the dot notation access back up in the view. Optionally, you can pass btn_name and btn_link to include a button in the last column of the table. In the case of managing memberships, there is a button that says 'manage' in each row in the 'Subscription' column. This links to the membership#show where a member will be able to do things like change their subscription, check payments, etc. This is where I'm planning to start implementing UI for the Stripe API.