Add SafeSend module
I created a module to wrap public_send in so I wasn't calling public_send directly from a view. Call me The Great Wall of Ruby. The Ruby Dome. Knowing me, you should probably refer to this as more of a Maginot Line situation, but we'll burn that bridge when we come to it.

The module has some whitelist/blacklists of methods and only calls public_send if the attempted method is whitelisted. Mmmm. Cybersecurity or whatever.