Add the ability for resolve_attr method to handle procs
I finally got the subscription column to render correctly on sites#show in reference to each visitor. Maybe I was over thinking it, but it was actually pretty difficult. I finally got it to work by allowing the attrs list you pass to the partial to handle an attr that's a lambda. I forget if I mentioned in a previous commit that I made a SafeSend module to wrap public_send in some whitelist/blacklist stuff for allowed methods, but I did that. I tried to recreate the same idea for the lambdas by defining a SafeLambda class in safe_send.rb, but I haven't quite gotten it to work yet. Something about the args not being passed correctly, but I think it will be cool once it works, and it will be much more secure because randoms can't pass a lambda to run which is obviously way worse thanwhen I was only allowing strings that can run methods on the base object. Oops. Is all this technically metaprogramming? *he asks into the void*